메뉴 열리는 QA 반영
